Fixes a leak of ejabberd_receiver processes.
When a (non-frontend) socket module without any custom receiver fails to
start, the newly created ejabberd_receiver process needs to be properly
closed.

This patch fixes this problem, as well as provides a solution for
stopping 'dead' (as in with dead sockets/ports) ejabberd_receiver
processes, by sending them the message 'check_socket'.