Include client namespace in CAPs queries (EJAB-1411)
When constructing an IQ record, if no namespace is specified
for the IQ,  exmpp treat that literally as "no namespace" and
so put xmlns="" (it do not interpret it as "no change" and let
the parent namespace apply).
So it is neccesarly to especify the jabber:client namespace.

See EXMPP-53.