Drop useless attrributes from user/group filter queries.
When running an LDAP query with gfilter, we don't need to fetch "group_attr".
It used to be needed in old version of this code which fetched information for
ALL groups in one query, so it needed an attribute to distinguish groups by.
Now that we query for one group info at a time, we no longer need this.

The same applies to ufilter and the "user_uid" attribute. Moreover this
attribute is not used anywhere else, so I drop parsing and storing of the
"ldap_useruid" configuration parameter completely.