Improve consistency by renaming uid*->member_attr*.
A few configuration parameters are called ldap_memberattr*, but internally in
the module their values are kept in variables/fields called uid*. This
inconsistency requires extra effort to translate in one's mind when writing
and updating documentation.

Rename some fields and variables from uid* to member_attr* to remove this
historical inconsistency.  None of these changes are visible to the user or
administrator.