Fix invalid iq reply handling in ejabberd_local
The problem was a post-binarisation leftover which caused mod_ping to
always assume no reply to a ping was received, which wasn't true.
The reply could be received,
but the iq reply handler was looked for under
a different key (a binary taken from 'id' attribute),
while the real key was an ordinary string/list.