Updated purge offline filter
When the pubsub nodes publish_model is open, then anyone could have
items that need to be purged, even if the person going offline is
not affiliated with the node.  The previous guard expression aborts
purge offline incorrectly in this case.  Instead, continue with the
purge if the user is affiliated with the node or if the node has an
open publish_model.

Change-Id: I2badd8dd3f6813cff91bb9915dc7e563d1d88c1b