Use -set-cookie when running erl in ejabberdctl.template.
This patch converts the ejabberdctl.template to use erl's
-set-cookie option when an .erlang.cookie file is found in
$SPOOL_DIR. This will help for installations where ejabberd might
be started by one user, but ejabberdctl might be used by a
different user to manage it. For example, in Fedora ejabberd is
launched by systemd as the ejabberd user, but users might want to
use "sudo ejabberdctl" to manage the node.