Fix for network.target dependency
Since the network.target is actually a requirement, it should be stated in "Requires=" as well. If missed, the system (tested on CentOS7) will attempt to stop the service after networking stops which doesn't work and causes unnecessary delay with a default (typically 90sec) timeout. Please note that the "After=" statement has to be there as well.