Add 'xml' format to mod_muc_log
This format will write the MUC messages into an XML file. The file will
contain stanzas (<message> for the messages, <presence> for joining, leaving
and nick changes and <iq type='set'> for root configuration changes).
Every stanzas will contain a <address> element with the JID of the user
sending the message and a <delay xmlns='urn:xmpp:delay'> element with the
time of the stanza. All messages for a day will be in a
<muclog xmlns='https://ejabberd.im/muclog'> element.