Set 'max_fsm_queue' default value (back) to 10000
The default value for the 'max_fsm_queue' option was set to 10000 in
commit 79685da90b514a49f9c412e31dccc17be9dd8680, and that value is still
documented to be the default.  It was (probably unintentionally) changed
to 5000 in commit 03de853e4fdcf852ae75a86922c08bb1a0950e6d.

It makes sense to keep it larger than the value of mod_stream_mgmt's
'max_ack_queue' option.